refactor: reorganize OCR module structure and consolidate functionality
This commit significantly improves the organization of OCR-related code by:

## Major Changes
- Move all OCR synchronous implementations from _multiprocessing to _ocr/_sync.py
- Move OCR process pool functionality to _ocr/_pool.py
- Consolidate ProcessPoolManager into _utils/_process_pool.py
- Update _multiprocessing module to focus only on GMFT isolation

## Module Reorganization
- Created comprehensive _ocr/_sync.py with all sync OCR implementations
- Created _ocr/_pool.py for process pool OCR functionality
- Updated _ocr/__init__.py to export all OCR functions consistently
- Streamlined _multiprocessing to handle only GMFT-related tasks

## Import Updates
- Updated all extractors to use new OCR module imports
- Fixed all test imports to reference correct module paths
- Updated Claude Code MCP server documentation

## Benefits
- Clear separation of concerns between OCR and multiprocessing
- Better modularity with centralized OCR functionality
- Maintained backward compatibility while improving structure
- All tests passing with new organization
