+++ kreuzberg/_multiprocessing/__init__.py
@@ -1,6 +1,5 @@
 """Multiprocessing utilities for kreuzberg."""
 
-from .process_manager import ProcessPoolManager
-from .tesseract_pool import TesseractProcessPool
+from .gmft_isolated import extract_tables_isolated, extract_tables_isolated_async
 
-__all__ = ["ProcessPoolManager", "TesseractProcessPool"]
+__all__ = ["extract_tables_isolated", "extract_tables_isolated_async"]
