+++ tests/multiprocessing/sync_easyocr_test.py
@@ -11,14 +11,14 @@ import pytest
 from PIL import Image
 
 from kreuzberg._mime_types import PLAIN_TEXT_MIME_TYPE
-from kreuzberg._multiprocessing.sync_easyocr import (
-    _get_easyocr_instance,
-    process_batch_images_sync_pure,
-    process_batch_images_threaded,
-    process_image_bytes_sync_pure,
-    process_image_sync_pure,
-)
 from kreuzberg._ocr._easyocr import EasyOCRConfig
+from kreuzberg._ocr._sync import (
+    _get_easyocr_instance,
+    process_batch_images_sync,
+    process_batch_images_threaded,
+    process_image_bytes_easyocr_sync,
+    process_image_easyocr_sync,
+)
 from kreuzberg._types import ExtractionResult
 from kreuzberg.exceptions import MissingDependencyError, OCRError
 
@@ -132,6 +132,14 @@ def test_get_easyocr_instance_legacy_use_gpu_config() -> None:
     config = Mock()
     config.use_gpu = True  # Legacy attribute
     config.language = "en"  # Add language attribute
+    # Set default None values for all optional attributes
+    config.model_storage_directory = None
+    config.user_network_directory = None
+    config.recog_network = None
+    config.detector = None
+    config.recognizer = None
+    config.quantize = None
+    config.cudnn_benchmark = None
     # Don't have device attribute
 
     mock_easyocr_module = Mock()
@@ -228,7 +236,7 @@ def test_get_easyocr_instance_advanced_config() -> None:
             mock_easyocr_module.Reader.assert_called_once_with(**expected_call_kwargs)
 
 
-def test_process_image_sync_pure_success(sample_image_path: str, mock_easyocr: Mock) -> None:
+def test_process_image_easyocr_sync_success(sample_image_path: str, mock_easyocr: Mock) -> None:
     """Test successful image processing with sync pure implementation."""
     # Mock EasyOCR results
     mock_results = [
@@ -238,7 +246,7 @@ def test_process_image_sync_pure_success(sample_image_path: str, mock_easyocr: M
     mock_easyocr.readtext.return_value = mock_results
 
     config = EasyOCRConfig()
-    result = process_image_sync_pure(sample_image_path, config)
+    result = process_image_easyocr_sync(sample_image_path, config)
 
     assert isinstance(result, ExtractionResult)
     assert result.content == "Hello\nWorld"
@@ -249,12 +257,12 @@ def test_process_image_sync_pure_success(sample_image_path: str, mock_easyocr: M
     mock_easyocr.readtext.assert_called_once()
 
 
-def test_process_image_sync_pure_no_results(sample_image_path: str, mock_easyocr: Mock) -> None:
+def test_process_image_easyocr_sync_no_results(sample_image_path: str, mock_easyocr: Mock) -> None:
     """Test image processing when EasyOCR returns no results."""
     mock_easyocr.readtext.return_value = []
 
     config = EasyOCRConfig()
-    result = process_image_sync_pure(sample_image_path, config)
+    result = process_image_easyocr_sync(sample_image_path, config)
 
     assert isinstance(result, ExtractionResult)
     assert result.content == ""
@@ -263,22 +271,22 @@ def test_process_image_sync_pure_no_results(sample_image_path: str, mock_easyocr
     assert result.chunks == []
 
 
-def test_process_image_sync_pure_default_config(sample_image_path: str, mock_easyocr: Mock) -> None:
+def test_process_image_easyocr_sync_default_config(sample_image_path: str, mock_easyocr: Mock) -> None:
     """Test image processing with default config (None)."""
     mock_easyocr.readtext.return_value = [([[0, 0], [100, 0], [100, 20], [0, 20]], "Test", 0.9)]
 
-    result = process_image_sync_pure(sample_image_path, None)
+    result = process_image_easyocr_sync(sample_image_path, None)
 
     assert isinstance(result, ExtractionResult)
     assert result.content == "Test"
     assert result.metadata == {"confidence": 0.9}
 
 
-def test_process_image_sync_pure_detail_false(sample_image_path: str, mock_easyocr: Mock) -> None:
+def test_process_image_easyocr_sync_detail_false(sample_image_path: str, mock_easyocr: Mock) -> None:
     """Test image processing with detail=False config."""
     mock_easyocr.readtext.return_value = ["Hello", "World"]
 
-    # Create a simple mock config with detail=False
+    # Create a config with all needed attributes
     config = Mock()
     config.decoder = "greedy"
     config.beam_width = 5
@@ -296,64 +304,69 @@ def test_process_image_sync_pure_detail_false(sample_image_path: str, mock_easyo
     config.add_margin = 0.1
     config.x_ths = 1.0
     config.y_ths = 0.5
+    config.detail = False  # This is what we want to test
+    config.device = None
+    config.language = "en"
+    # Set all optional attributes to None
+    config.model_storage_directory = None
+    config.user_network_directory = None
+    config.recog_network = None
+    config.detector = None
+    config.recognizer = None
+    config.quantize = None
+    config.cudnn_benchmark = None
+    config.batch_size = 1
+    config.workers = 0
+    config.allowlist = None
+    config.blocklist = None
+    config.paragraph = False
 
-    # Mock getattr to return detail=False for lines 127-128 coverage
-    def mock_getattr(obj: Any, attr: str, default: Any = None) -> Any:
-        if attr == "detail":
-            return False  # This triggers lines 127-128
-        if hasattr(config, attr):
-            return getattr(config, attr)
-        return default
+    result = process_image_easyocr_sync(sample_image_path, config)
 
-    with patch("builtins.getattr", side_effect=mock_getattr):
-        result = process_image_sync_pure(sample_image_path, config)
-
-        assert isinstance(result, ExtractionResult)
-        assert result.content == "Hello\nWorld"
-        assert result.metadata == {"confidence": 1.0}  # Default confidence when detail=False
+    assert isinstance(result, ExtractionResult)
+    assert result.content == "Hello\nWorld"
+    assert result.metadata == {"confidence": 1.0}  # Default confidence when detail=False
 
 
-def test_process_image_sync_pure_short_result(sample_image_path: str, mock_easyocr: Mock) -> None:
+def test_process_image_easyocr_sync_short_result(sample_image_path: str, mock_easyocr: Mock) -> None:
     """Test image processing with short result (missing confidence)."""
     # Result with only 2 elements (no confidence)
     mock_results = [([[0, 0], [100, 0], [100, 20], [0, 20]], "Hello")]
     mock_easyocr.readtext.return_value = mock_results
 
     config = EasyOCRConfig()
-    result = process_image_sync_pure(sample_image_path, config)
+    result = process_image_easyocr_sync(sample_image_path, config)
 
     assert isinstance(result, ExtractionResult)
     assert result.content == "Hello"
     assert result.metadata == {"confidence": 1.0}  # Default confidence
 
 
-def test_process_image_sync_pure_invalid_result(sample_image_path: str, mock_easyocr: Mock) -> None:
+def test_process_image_easyocr_sync_invalid_result(sample_image_path: str, mock_easyocr: Mock) -> None:
     """Test image processing with invalid result structure."""
     # Result with only 1 element (invalid)
     mock_results = [([[0, 0], [100, 0], [100, 20], [0, 20]],)]
     mock_easyocr.readtext.return_value = mock_results
 
     config = EasyOCRConfig()
-    result = process_image_sync_pure(sample_image_path, config)
+    result = process_image_easyocr_sync(sample_image_path, config)
 
     assert isinstance(result, ExtractionResult)
     assert result.content == ""
     assert result.metadata == {}
 
 
-def test_process_image_sync_pure_exception(sample_image_path: str) -> None:
+def test_process_image_easyocr_sync_exception(sample_image_path: str) -> None:
     """Test OCRError is raised when EasyOCR processing fails."""
     config = EasyOCRConfig()
 
-    with patch(
-        "kreuzberg._multiprocessing.sync_easyocr._get_easyocr_instance", side_effect=Exception("EasyOCR failed")
-    ):
+    with patch("kreuzberg._ocr._sync._get_easyocr_instance", side_effect=Exception("EasyOCR failed")):
         with pytest.raises(OCRError, match="EasyOCR processing failed"):
-            process_image_sync_pure(sample_image_path, config)
+            process_image_easyocr_sync(sample_image_path, config)
 
 
 @pytest.mark.skip("Complex mocking leads to issues")
-def test_process_image_sync_pure_advanced_config(sample_image_path: str, mock_easyocr: Mock) -> None:
+def test_process_image_easyocr_sync_advanced_config(sample_image_path: str, mock_easyocr: Mock) -> None:
     """Test image processing with advanced EasyOCR configuration."""
     mock_easyocr.readtext.return_value = [([[0, 0], [100, 0], [100, 20], [0, 20]], "Test", 0.9)]
 
@@ -397,7 +410,7 @@ def test_process_image_sync_pure_advanced_config(sample_image_path: str, mock_ea
 
         mock_getattr.side_effect = getattr_side_effect
 
-        result = process_image_sync_pure(sample_image_path, config)
+        result = process_image_easyocr_sync(sample_image_path, config)
 
         assert isinstance(result, ExtractionResult)
         assert result.content == "Test"
@@ -431,27 +444,27 @@ def test_process_image_sync_pure_advanced_config(sample_image_path: str, mock_ea
         assert kwargs["y_ths"] == 0.5
 
 
-def test_process_image_bytes_sync_pure_success(sample_image_bytes: bytes, mock_easyocr: Mock) -> None:
+def test_process_image_bytes_easyocr_sync_success(sample_image_bytes: bytes, mock_easyocr: Mock) -> None:
     """Test successful image bytes processing."""
     mock_easyocr.readtext.return_value = [([[0, 0], [100, 0], [100, 20], [0, 20]], "Bytes Test", 0.85)]
 
     config = EasyOCRConfig()
-    result = process_image_bytes_sync_pure(sample_image_bytes, config)
+    result = process_image_bytes_easyocr_sync(sample_image_bytes, config)
 
     assert isinstance(result, ExtractionResult)
     assert result.content == "Bytes Test"
     assert result.metadata == {"confidence": 0.85}
 
 
-def test_process_image_bytes_sync_pure_cleanup() -> None:
+def test_process_image_bytes_easyocr_sync_cleanup() -> None:
     """Test that temporary files are cleaned up after processing."""
     sample_bytes = b"fake image data"
 
-    with patch("kreuzberg._multiprocessing.sync_easyocr.Image.open") as mock_open:
+    with patch("kreuzberg._ocr._sync.Image.open") as mock_open:
         mock_image = Mock()
         mock_open.return_value.__enter__.return_value = mock_image
 
-        with patch("kreuzberg._multiprocessing.sync_easyocr.process_image_sync_pure") as mock_process:
+        with patch("kreuzberg._ocr._sync.process_image_easyocr_sync") as mock_process:
             mock_process.return_value = ExtractionResult(
                 content="test", mime_type=PLAIN_TEXT_MIME_TYPE, metadata={}, chunks=[]
             )
@@ -470,7 +483,7 @@ def test_process_image_bytes_sync_pure_cleanup() -> None:
                 return tf
 
             with patch("tempfile.NamedTemporaryFile", side_effect=track_tempfile):
-                result = process_image_bytes_sync_pure(sample_bytes, None)
+                result = process_image_bytes_easyocr_sync(sample_bytes, None)
 
                 assert isinstance(result, ExtractionResult)
                 assert result.content == "test"
@@ -480,7 +493,7 @@ def test_process_image_bytes_sync_pure_cleanup() -> None:
                     assert not Path(temp_file).exists()
 
 
-def test_process_batch_images_sync_pure_success(mock_easyocr: Mock) -> None:
+def test_process_batch_images_sync_success(mock_easyocr: Mock) -> None:
     """Test successful batch processing with sync pure implementation."""
     # Create temporary image files
     image_paths = []
@@ -496,7 +509,7 @@ def test_process_batch_images_sync_pure_success(mock_easyocr: Mock) -> None:
         mock_easyocr.readtext.side_effect = side_effect_results
 
         config = EasyOCRConfig()
-        results = process_batch_images_sync_pure(cast("list[str | Path]", image_paths), config)
+        results = process_batch_images_sync(cast("list[str | Path]", image_paths), config, backend="easyocr")
 
         assert len(results) == 3
         for i, result in enumerate(results):
@@ -510,9 +523,9 @@ def test_process_batch_images_sync_pure_success(mock_easyocr: Mock) -> None:
             Path(path).unlink(missing_ok=True)
 
 
-def test_process_batch_images_sync_pure_empty_list(mock_easyocr: Mock) -> None:
+def test_process_batch_images_sync_empty_list(mock_easyocr: Mock) -> None:
     """Test batch processing with empty image list."""
-    results = process_batch_images_sync_pure([], None)
+    results = process_batch_images_sync([], None, backend="easyocr")
     assert results == []
 
 
@@ -532,7 +545,9 @@ def test_process_batch_images_threaded_success(mock_easyocr: Mock) -> None:
         mock_easyocr.readtext.side_effect = side_effect_results
 
         config = EasyOCRConfig()
-        results = process_batch_images_threaded(cast("list[str | Path]", image_paths), config, max_workers=2)
+        results = process_batch_images_threaded(
+            cast("list[str | Path]", image_paths), config, backend="easyocr", max_workers=2
+        )
 
         assert len(results) == 3
         # Results should be in same order as input
@@ -567,7 +582,9 @@ def test_process_batch_images_threaded_with_errors(mock_easyocr: Mock) -> None:
         mock_easyocr.readtext.side_effect = side_effect
 
         config = EasyOCRConfig()
-        results = process_batch_images_threaded(cast("list[str | Path]", image_paths), config, max_workers=2)
+        results = process_batch_images_threaded(
+            cast("list[str | Path]", image_paths), config, backend="easyocr", max_workers=2
+        )
 
         assert len(results) == 3
         assert isinstance(results[0], ExtractionResult)
@@ -591,7 +608,7 @@ def test_process_batch_images_threaded_default_workers(mock_easyocr: Mock) -> No
     """Test threaded batch processing with default worker count."""
     image_paths = ["image1.png", "image2.png"]  # Fake paths for this test
 
-    with patch("kreuzberg._multiprocessing.sync_easyocr.process_image_sync_pure") as mock_process:
+    with patch("kreuzberg._ocr._sync.process_image_easyocr_sync") as mock_process:
         mock_process.return_value = ExtractionResult(
             content="test", mime_type=PLAIN_TEXT_MIME_TYPE, metadata={}, chunks=[]
         )
@@ -607,5 +624,5 @@ def test_process_batch_images_threaded_default_workers(mock_easyocr: Mock) -> No
 
 def test_process_batch_images_threaded_empty_list() -> None:
     """Test threaded batch processing with empty image list."""
-    results = process_batch_images_threaded([], None, max_workers=2)
+    results = process_batch_images_threaded([], None, backend="tesseract", max_workers=2)
     assert results == []
