+++ tests/multiprocessing/tesseract_pool_test.py
@@ -9,7 +9,7 @@ from unittest.mock import patch
 import pytest
 from PIL import Image
 
-from kreuzberg._multiprocessing.tesseract_pool import (
+from kreuzberg._ocr._pool import (
     TesseractProcessPool,
     _process_image_bytes_with_tesseract,
     _process_image_with_tesseract,
